Fix duplicated files detection

+++ remove_special_chars.py
@@ -4,7 +4,7 @@ import unidecode
 import sys
 
 
-def remove_chars(file_name):
+def remove_chars(file_name, suffix=''):
     """
     Removes special characters and change diacritics by its natural char.
 
@@ -17,6 +17,7 @@ def remove_chars(file_name):
     # is the file extension
     file_no_extension = ''.join(file_name_split[:-1]) if len(file_name_split) > 1 else file_name_split[0]
     new_str = re.sub('[^a-zA-Z0-9]', '', unidecode.unidecode(file_no_extension))
+    new_str = '{}{}'.format(new_str, str(suffix) if suffix else '')
     return '.'.join([new_str, file_name_split[-1]]) if len(file_name_split) > 1 else new_str
 
 def recursive_change_file_name(parent_dir, file_name=''):
@@ -24,12 +25,21 @@ def recursive_change_file_name(parent_dir, file_name=''):
     Recursively change the name of the directories
     """
     file_dir = os.path.join(parent_dir, file_name)
-    print(file_dir)
+    # print(file_dir)
     if file_name:
-        new_dir_name = remove_chars(file_name)
-        new_file_dir = os.path.join(parent_dir, new_dir_name)
-        os.rename(file_dir, new_file_dir)
-        file_dir = new_file_dir
+        file_exists = True
+        suffix = 0
+        while file_exists:
+            new_dir_name = remove_chars(file_name, suffix)
+            new_file_dir = os.path.join(parent_dir, new_dir_name)
+            if os.path.isfile(new_file_dir) or os.path.isdir(new_file_dir):
+                suffix += 1
+                continue
+            else:
+                os.rename(file_dir, new_file_dir)
+                print(new_dir_name, new_file_dir)
+                file_exists = False
+                file_dir = new_file_dir
 
     try:
         files_list = os.listdir(file_dir)
